src/requirements.txt
mysql-connector-python==8.0.33
requests==2.31.0
deepdiff==6.3.0
loguru